requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Supabase
supabase==2.9.0

# Stripe
stripe==11.1.0

# Email
resend==2.5.1

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.3.0
httpx==0.27.0
